Create Targeted repeat macro
Alternative version of the default <<repeat>> macro (comes with its own version of <<stop>> -> <<st>>).

Syntax:

<<rep TimeSeconds '#TargetID(optional)' >>
Contents

<</rep>>

The default <<repeat>> creates a single span to which it appends the generated content (making it unfit to populate a grid for example). 
This version create a new numbered span on each iteration, if an #ID is supplied as argument the contents will be appended there, else in passage.

Credit to TheMadExile for the original code.